+++ src/test/java/store/aurora/order/service/OrderServiceImplTest.java
@@ -1,7 +1,6 @@
-package store.aurora.order.service.impl;
+package store.aurora.order.service;
 
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
