+++ src/test/java/store/aurora/order/service/ShipmentInformationServiceImplTest.java
@@ -1,4 +1,4 @@
-package store.aurora.order.service.impl;
+package store.aurora.order.service;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
