+++ src/test/java/store/aurora/order/service/ShipmentServiceImplTest.java
@@ -1,6 +1,5 @@
-package store.aurora.order.service.impl;
+package store.aurora.order.service;
 
-import org.hibernate.annotations.processing.SQL;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
